Refactor Environment,
- allow Extractor plugins (instead of only SourceExtractor)
- make_converter returns a callable directly
- add make_transform
